backend/pyproject.toml
[project]
name = "student-resume-portrait-backend"
version = "0.1.0"
description = "Student resume parsing and portrait system backend"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "sqlalchemy>=2.0.0",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.3.0",
  "python-multipart>=0.0.9",
  "celery>=5.4.0",
  "redis>=5.0.0",
  "python-docx>=1.1.2",
  "PyMuPDF>=1.24.0",
  "pillow>=10.0.0",
  "pgvector>=0.3.0",
  "httpx>=0.28.0",
  "pytest>=8.0.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
